SamplerBaseを削除、SamplerOptimizedに統合
