+++ src/SamplerOptimized.cpp
@@ -2,6 +2,16 @@
 #include <algorithm>
 #include <tables.h>
 
+Sample *Timbre::GetAppropriateSample(uint8_t noteNo, uint8_t velocity)
+{
+    for (MappedSample ms : samples)
+    {
+        if (ms.lowerNoteNo <= noteNo && noteNo <= ms.upperNoteNo && ms.lowerVelocity <= velocity && velocity <= ms.upperVelocity)
+            return ms.sample;
+    }
+    return nullptr;
+}
+
 void SamplerOptimized::SetTimbre(uint8_t channel, Timbre *t)
 {
     if(channel < CH_COUNT) channels[channel].SetTimbre(t);
